# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swagger documentation fo metallplace service",
        "title": "Metallplace API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Ivan Demchuk",
            "url": "http://www.swagger.io/support",
            "email": "is.demchuk@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/getMonthlyAvgFeed": {
            "post": {
                "description": "get n last values of specific property of specific material",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "value"
                ],
                "summary": "Get last values",
                "parameters": [
                    {
                        "description": "query params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.AvgPriceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.AvgPriceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/getReport/{type}/{date}": {
            "get": {
                "description": "get generated report in docx format. Type could be ether \"weekly\" or \"monthly\". Date of report in format YYYY-MM-DD",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "report"
                ],
                "summary": "Get report file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/getWeeklyAvgFeed": {
            "post": {
                "description": "returns weekly feed of average values during the week",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "value"
                ],
                "summary": "Get weekly feed",
                "parameters": [
                    {
                        "description": "query params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.AvgPriceRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.AvgPriceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.AvgPriceRequest": {
            "type": "object",
            "properties": {
                "finish": {
                    "type": "string",
                    "format": "string",
                    "example": "2020-01-01"
                },
                "material_source_id": {
                    "type": "integer",
                    "format": "int64",
                    "example": 2
                },
                "property_id": {
                    "type": "integer",
                    "format": "int64",
                    "example": 1
                },
                "start": {
                    "type": "string",
                    "format": "string",
                    "example": "2019-01-01"
                }
            }
        },
        "handler.AvgPriceResponse": {
            "type": "object",
            "properties": {
                "prev_price": {
                    "type": "number",
                    "format": "float64",
                    "example": 42.55
                },
                "price_feed": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Price"
                    }
                }
            }
        },
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.Price": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string",
                    "format": "string",
                    "example": "2022-03-23"
                },
                "value": {
                    "type": "number",
                    "format": "float64",
                    "example": 12.97
                }
            }
        },
...
    },
...
}